+++ config/.eslintrc.vue.js
@@ -1,0 +1,10 @@
+const base = require('./.eslint.base')
+module.exports = {
+  ...base,
+  extends: ['plugin:import/recommended', 'plugin:import/typescript'],
+  parser: 'vue-eslint-parser',
+  parserOptions: {
+    ...base.parserOptions,
+    parser: '@typescript-eslint/parser'
+  }
+}