+++ examples/vue-apollo/src/pages/AboutPage.vue
@@ -1,0 +1,7 @@
+<template>
+  <div className="d-flex align-center flex-column">
+    <v-card width="400">
+      <v-card-text> About </v-card-text>
+    </v-card>
+  </div>
+</template>