+++ packages/hasura-auth-js/tests/sign-out.test.ts
@@ -1,10 +1,11 @@
+import { USER_UNAUTHENTICATED } from '@nhost/core'
 import faker from 'faker'
 import { describe, expect, it } from 'vitest'
 
 import { auth, signUpAndInUser } from './helpers'
 
 describe('sign-out', () => {
-  it('sign in user with email and password', async () => {
+  it('should sign out', async () => {
     const email = faker.internet.email().toLocaleLowerCase()
     const password = faker.internet.password(8)
 
@@ -12,8 +13,28 @@ describe('sign-out', () => {
 
     expect(auth.getSession()).toBeTruthy()
 
-    await auth.signOut()
+    const { error } = await auth.signOut()
+    expect(error).toBeNull()
+    expect(auth.getSession()).toBeNull()
+  })
+
+  it('should sign out from all devices', async () => {
+    const email = faker.internet.email().toLocaleLowerCase()
+    const password = faker.internet.password(8)
+
+    await signUpAndInUser({ email, password })
+
+    expect(auth.getSession()).toBeTruthy()
+
+    const { error } = await auth.signOut({ all: true })
+    expect(error).toBeNull()
 
     expect(auth.getSession()).toBeNull()
+  })
+
+  it('should not be possible to signout when not authenticated', async () => {
+    const { error } = await auth.signOut()
+
+    expect(error).toEqual(USER_UNAUTHENTICATED)
   })
 })