+++ packages/vue/.eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: '../../config/.eslintrc.vue.js'
+}