+++ packages/vue/src/useUserEmail.ts
@@ -1,0 +1,22 @@
+import { useSelector } from '@xstate/vue'
+
+import { useAuthInterpreter } from './useAuthInterpreter'
+
+/**
+ * Use the composbale `useUserEmail` to get the email of the user.
+ *
+ * @example
+ * ```tsx
+ * const userEmail = useUserEmail();
+ * ```
+ *
+ * @docs https://docs.nhost.io/reference/vue/use-user-email
+ */
+export const useUserEmail = () => {
+  const service = useAuthInterpreter()
+  return useSelector(
+    service.value,
+    (state) => state.context.user?.email,
+    (a, b) => a === b
+  )
+}