+++ packages/hasura-auth-js/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @nhost/hasura-auth-js
 
+## 1.1.8
+
+### Patch Changes
+
+- Updated dependencies [6c423394]
+  - @nhost/core@0.5.6
+
+## 1.1.7
+
+### Patch Changes
+
+- Updated dependencies [0b58894e]
+- Updated dependencies [c7a407f1]
+  - @nhost/core@0.5.5
+
 ## 1.1.6
 
 ### Patch Changes