+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @nhost/nextjs
 
+## 1.2.7
+
+### Patch Changes
+
+- Updated dependencies [6c423394]
+  - @nhost/core@0.5.6
+  - @nhost/react@0.7.7
+  - @nhost/nhost-js@1.1.13
+
+## 1.2.6
+
+### Patch Changes
+
+- Updated dependencies [0b58894e]
+- Updated dependencies [c7a407f1]
+  - @nhost/core@0.5.5
+  - @nhost/react@0.7.6
+  - @nhost/nhost-js@1.1.12
+
 ## 1.2.5
 
 ### Patch Changes