+++ packages/nhost-js/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nhost/nhost-js
 
+## 1.1.13
+
+### Patch Changes
+
+- @nhost/hasura-auth-js@1.1.8
+
+## 1.1.12
+
+### Patch Changes
+
+- @nhost/hasura-auth-js@1.1.7
+
 ## 1.1.11
 
 ### Patch Changes