+++ packages/react-apollo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nhost/react-apollo
 
+## 4.2.7
+
+### Patch Changes
+
+- @nhost/react@0.7.7
+- @nhost/apollo@0.5.6
+
+## 4.2.6
+
+### Patch Changes
+
+- @nhost/react@0.7.6
+- @nhost/apollo@0.5.5
+
 ## 4.2.5
 
 ### Patch Changes